catalog-service/requirements.txt
fastapi
requests
pydantic-settings

sqlalchemy
psycopg2-binary

uvicorn
